fixup! Noticable value for dummy implementation

+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/service/devices/garmin/VivoFit3Support.java
@@ -107,7 +107,7 @@ public class VivoFit3Support extends AbstractBTLEDeviceSupport {
 
 	private static class CRC16 implements Checksum {
 		private static short polynomial = (short) 0x8005;
-		private short acc = 0xABCD;
+		private short acc = (short) 0xABCD;
 
 		public void reset() {
 			acc = 0;
